+++ packages/guideline-blocks-settings/src/components/BlockItemWrapper/Toolbar/ToolbarButton/ToolbarButtonTooltip.tsx
@@ -1,0 +1,24 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+import { LegacyTooltip as Tooltip, TooltipPosition } from '@frontify/fondue';
+import { ReactElement, ReactNode } from 'react';
+
+type ToolbarButtonTooltipProps = {
+    content: ReactNode;
+    children: ReactElement;
+    open?: boolean;
+    disabled?: boolean;
+};
+
+export const ToolbarButtonTooltip = ({ open, content, children, disabled }: ToolbarButtonTooltipProps) => (
+    <Tooltip
+        withArrow
+        hoverDelay={0}
+        enterDelay={300}
+        open={open}
+        disabled={disabled}
+        position={TooltipPosition.Top}
+        content={<div>{content}</div>}
+        triggerElement={children}
+    />
+);
