+++ packages/guideline-blocks-settings/src/components/BlockItemWrapper/Toolbar/context/DragPreviewContext.ts
@@ -1,0 +1,7 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+import { createContext, useContext } from 'react';
+
+export const DragPreviewContext = createContext(false);
+
+export const useIsDragPreview = () => useContext(DragPreviewContext);
