+++ packages/guideline-blocks-settings/src/components/BlockItemWrapper/Toolbar/context/FlyoutCloseContext.ts
@@ -1,0 +1,7 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+import { createContext, useContext } from 'react';
+
+export const FlyoutCloseContext = createContext<() => void>(() => {});
+
+export const useCloseFlyoutContext = () => useContext(FlyoutCloseContext);
