+++ packages/guideline-blocks-settings/src/components/BlockItemWrapper/Toolbar/types.ts
@@ -1,11 +1,10 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
-import { type MenuItemStyle } from '@frontify/fondue';
-import { HTMLAttributes, ReactElement, ReactNode } from 'react';
+import { DragHandleToolbarButtonProps, FlyoutMenuItem, ToolbarButtonProps } from './ToolbarButton';
 
 export type ToolbarProps = {
     items: ToolbarItem[];
-    flyoutMenu: ToolbarFlyoutState & { items: FlyoutToolbarItem[][] };
+    flyoutMenu: ToolbarFlyoutState & { items: FlyoutMenuItem[][] };
     attachments: ToolbarFlyoutState & { isEnabled: boolean };
     isDragging?: boolean;
 };
@@ -15,37 +14,10 @@ export type ToolbarFlyoutState = {
     onOpenChange: (isOpen: boolean) => void;
 };
 
-type BaseToolbarItem = {
-    icon: JSX.Element;
-    tooltip?: string;
-};
+export type DraghandleToolbarItem = DragHandleToolbarButtonProps;
 
-export type DraghandleToolbarItem = BaseToolbarItem & {
-    draggableProps: Record<string, unknown>;
-    setActivatorNodeRef?: (node: HTMLElement | null) => void;
-};
+export type ButtonToolbarItem = ToolbarButtonProps;
 
-export type ButtonToolbarItem = BaseToolbarItem & {
-    onClick: () => void;
-};
+export type FlyoutToolbarItem = FlyoutMenuItem;
 
 export type ToolbarItem = DraghandleToolbarItem | ButtonToolbarItem;
-
-export type FlyoutToolbarItem = {
-    title: string;
-    onClick: () => void;
-    icon: JSX.Element;
-    style?: MenuItemStyle;
-};
-
-export type ToolbarButtonTooltipProps = {
-    content: ReactNode;
-    children: ReactElement;
-    open?: boolean;
-    disabled?: boolean;
-};
-
-export type BaseToolbarButtonProps = {
-    children: ReactNode;
-    forceActiveStyle?: boolean;
-} & HTMLAttributes<HTMLButtonElement>;
